Make sure that the suffix is always uppercase in the createLoader() method.
We need to make sure that the suffix is uppercase when comparing it in the
factory method. Therefore do not rely on an uppercase version of the suffix in
the createLoader() method as a parameter, but generate it in the method itself.

This method is normally only called by the ResourceLoader class, but for the
sake of code sanity it should still not rely on the caller...